requirements.txt
pypdf
langchain~=0.3.9
chromadb # vector db
pytest
boto3
langchain_community
anthropic~=0.40.0
lxml # html
transformers~=4.46.3
accelerate
openai~=1.56.0
Flask~=3.1.0
python-dotenv~=1.0.1
openai-whisper
bs4~=0.0.2
beautifulsoup4~=4.12.3
requests~=2.32.3